Expose a pairing channel API
