+++ README.md
@@ -7,6 +7,5 @@ working through the details.
 
 For now, you can try out a "mock" version of the crypto layer which uses
 the same message sequence as TLS1.3, but doesn't actually do any of the crypto.
-See the example integration in `./demo/tlschannel.js` and try it out live
-by loading `./demo/test_client.html` and `./demo/test_server.html` in
+Try it out live by loading `./demo/test_client.html` and `./demo/test_server.html` in
 parallel webpages.
