+++ src/index.js
@@ -4,224 +4,164 @@
 
 'use strict';
 
-// The top-level APIs offered by this library are `ClientConnection` and
-// `ServerConnection` classes.  They each take a callback to be used for
-// sending data to the remote peer, and operate like this:
-//
-//    conn = await ClientConnection.create(psk, pskId, async function send_data_to_server(data) {
-//      // application-specific sending logic here.
-//    })
-//
-//    // Send data to the server by calling `send`,
-//    // which will use the callback provided in the constructor.
-//    // A single `send()` by the application may result in multiple
-//    // invokations of the callback.
-//
-//    await conn.send('application-level data')
-//
-//    // When data is received from the server, push it into
-//    // the connection and let it return any decrypted app-level data.
-//    // There might not be any app-level data if it was protocol control message,
-//    // and the receipt of the data might trigger additional calls to the
-//    // send callback for protocol control purposes.
-//
-//    serverSocket.on('data', async encrypted_data => {
-//      const plaintext = await conn.recv(data)
-//      if (plaintext !== null) {
-//        do_something_with_app_level_data(plaintext)
-//      }
-//    })
-//
-//    // It's good practice to explicitly close the connection
-//    // when finished.  This will send a "closed" notification
-//    // to the server.
-//
-//    await conn.close()
-//
-// The `ServerConnection` API operates similarly; the distinction is mainly
-// in which side is expected to send vs receieve during the protocol handshake.
+// A wrapper that combines a WebSocket to the channelserver
+// with a TLS Connection for encryption.
 
 import {
-  assert,
-  assertIsBytes,
+  InsecureClientConnection,
+  InsecureServerConnection,
+} from './tlsconnection.js';
+
+import {
   bytesToHex,
   hexToBytes,
   bytesToUtf8,
   utf8ToBytes,
-  noop,
 } from './utils.js';
 
-import {
-  getRandomBytes
-} from './crypto.js';
+// Copy pasta from https://developer.mozilla.org/en-US/docs/Web/API/EventTarget.
+var EventTarget = function() {
+  this.listeners = {};
+};
 
-import * as STATE from './states.js';
+EventTarget.prototype.listeners = null;
+EventTarget.prototype.addEventListener = function(type, callback) {
+  if (!(type in this.listeners)) {
+    this.listeners[type] = [];
+  }
+  this.listeners[type].push(callback);
+};
 
-import {
-  readHandshakeMessage,
-} from './messages.js';
+EventTarget.prototype.removeEventListener = function(type, callback) {
+  if (!(type in this.listeners)) {
+    return;
+  }
+  var stack = this.listeners[type];
+  for (var i = 0, l = stack.length; i < l; i++) {
+    if (stack[i] === callback) {
+      stack.splice(i, 1);
+      return;
+    }
+  }
+};
 
-import {
-  RecordSender,
-  RecordReceiver,
-  RECORD_TYPE,
-} from './recordlayer.js';
+EventTarget.prototype.dispatchEvent = function(event) {
+  if (!(event.type in this.listeners)) {
+    return true;
+  }
+  var stack = this.listeners[event.type].slice();
 
-// !!!!!!!
-// !!
-// !!   N.B. We have not yet implemented the actual encryption bits!
-// !!
-// !!!!!!!
-//
-// This first version of the code uses the same "framework" as we would use for
-// implementing TLS1.3, including the basic sequence of messages between client and
-// server, and the use of binary encoding in an ArrayBuffer for data processing.
-// Using an implementation with proper crypto should feel identical to using
-// this mock version, except it won't have "Insecure" in the class name...
+  for (var i = 0, l = stack.length; i < l; i++) {
+    stack[i].call(this, event);
+  }
+  return !event.defaultPrevented;
+};
 
-class InsecureConnection {
-  constructor(psk, pskId, sendCallback, randomSalt) {
-    this.psk = assertIsBytes(psk);
-    this.pskId = assertIsBytes(pskId);
-    this.sendCallback = sendCallback;
-    this.randomSalt = randomSalt;
-    this._state = new STATE.UNINITIALIZED(this);
-    this._pendingApplicationData = [];
-    this._recordSender = new RecordSender();
-    this._recordReceiver = new RecordReceiver();
-    this._lastPromise = Promise.resolve();
+export class InsecurePairingChannel extends EventTarget {
+  constructor(channelId, channelKey, socket, tlsConnection) {
+    super();
+    this._channelId = channelId;
+    this._channelKey = channelKey;
+    this._socket = socket;
+    this._tlsConnection = tlsConnection;
+    this._setupListeners();
   }
 
-  // Subclasses will override this with some async initialization logic.
-  static async create(psk, pskId, sendCallback, randomSalt = null) {
-    randomSalt = randomSalt === null ? await getRandomBytes(32) : randomSalt;
-    return new this(psk, pskId, sendCallback, randomSalt);
+  /**
+   * Create a new pairing channel.
+   *
+   * @returns Promise<InsecurePairingChannel>
+   */
+  static create(channelServerURI, {WebSocket} = window) {
+    const channelKey = crypto.getRandomValues(new Uint8Array(32));
+    return this._makePairingChannel(WebSocket, channelServerURI, InsecureServerConnection, channelKey);
   }
 
-  // These are the three public API methods that
-  // consumers can use to connunicate over TLS.
+  /**
+   * Connect to an existing pairing channel.
+   *
+   * `channelServerURI` must be the full URI and therefore should contain the channel ID.
+   * @returns Promise<InsecurePairingChannel>
+   */
+  static connect(channelServerURI, channelKey, {WebSocket} = window) {
+    return this._makePairingChannel(WebSocket, channelServerURI, InsecureClientConnection, channelKey);
+  }
 
-  async send(data) {
-    assertIsBytes(data);
-    await this._synchronized(async () => {
-      await this._state.sendApplicationData(data);
+  static _makePairingChannel(WebSocket, wsUri, TlsConnection, psk) {
+    const socket = new WebSocket(wsUri);
+    return new Promise((resolve, reject) => {
+      const stopListening = () => {
+        socket.removeEventListener("error", onConnectionError);
+        socket.removeEventListener("message", onFirstMessage);
+      };
+      const onConnectionError = async () => {
+        stopListening();
+        reject(new Error("Error while creating the pairing channel"));
+      };
+      const onFirstMessage = async event => {
+        stopListening();
+        try {
+          const {channelid: channelId} = JSON.parse(event.data);
+          const pskId = utf8ToBytes(channelId);
+          const tlsConnection = await TlsConnection.create(psk, pskId, data => {
+            // To send data over the websocket, it needs to be encoded as a safe string.
+            socket.send(bytesToHex(data));
+          });
+          const instance = new this(channelId, psk, socket, tlsConnection);
+          resolve(instance);
+        } catch (err) {
+          reject(err);
+        }
+      };
+      socket.addEventListener("error", onConnectionError);
+      socket.addEventListener("message", onFirstMessage);
     });
   }
 
-  async recv(data) {
-    assertIsBytes(data);
-    return await this._synchronized(async () => {
-      const [type, buf] = await this._recordReceiver.recv(data);
-      return await this._dispatchIncomingRecord(type, buf);
+  _setupListeners() {
+    this._socket.addEventListener("message", async event => {
+      let {message: payload, sender} = JSON.parse(event.data);
+      payload = await this._tlsConnection.recv(hexToBytes(payload));
+      if (payload !== null) {
+        const {message, data} = JSON.parse(bytesToUtf8(payload));
+        // Assume the application wants to deal with strings,
+        // not Uint8Array buffers as returned by the TLS layer.
+        this.dispatchEvent({
+          type: "message",
+          sender,
+          message,
+          data,
+        });
+      }
     });
+    // Relay the other events.
+    this._socket.addEventListener("error", this.dispatchEvent);
+    this._socket.addEventListener("close", this.dispatchEvent);
+  }
+
+  send(message, data = {}) {
+    const envelope = JSON.stringify({
+      message,
+      data,
+    });
+    this._tlsConnection.send(utf8ToBytes(envelope));
   }
 
   async close() {
-    await this._synchronized(async () => {
-      await this._state.close();
-    });
+    await this._tlsConnection.close();
+    this._tlsConnection = null;
+    this.socket.close();
   }
 
-  // Ensure that async functions execute one at a time,
-  // by waiting for the previous call to `_synchronized()` to complete
-  // before starting a new one.  This helps ensure that we complete
-  // one state-machine transition before starting to do the next.
-
-  _synchronized(cb) {
-    const nextPromise = this._lastPromise.then(() => {
-      return cb();
-    }).catch(async err => {
-      // All errors immediately put us in a terminal "error" state.
-      await this._transition(STATE.ERROR, err);
-      throw err;
-    });
-    // We don't want to hold on to the return value or error,
-    // just synchronize on the fact that it completed.
-    this._lastPromise = nextPromise.then(noop, noop);
-    return nextPromise;
+  get closed() {
+    return this._socket.readyState == 3;
   }
 
-  // This drives internal transition of the state-machine,
-  // ensuring that the new state is properly initialized.
-
-  async _transition(State, ...args) {
-    this._state = new State(this);
-    await this._state.initialize(...args);
+  get channelId() {
+    return this._channelId;
   }
 
-  // These are helpers to allow the state to add data to the next outgoing record.
-
-  async _writeApplicationData(bytes) {
-    await this._recordSender.withBufferWriter(RECORD_TYPE.APPLICATION_DATA, async buf => {
-      await buf.writeBytes(bytes);
-    });
-  }
-
-  async _writeHandshakeMessage(msg) {
-    await this._recordSender.withBufferWriter(RECORD_TYPE.HANDSHAKE, async buf => {
-      // XXX TODO: remember buffer position here.
-      await msg.write(buf);
-      // XXX TODO: read back to prev buffer position, add it to the transcript hash.
-    });
-  }
-
-  async _flushOutgoingRecord() {
-    const record = await this._recordSender.flush();
-    await this.sendCallback(record);
-  }
-
-  // This is a helper for handling incoming records.
-
-  async _dispatchIncomingRecord(type, buf) {
-    switch (type) {
-      case RECORD_TYPE.ALERT:
-        // XXX TODO: implement alert records for communicating errors.
-        throw new Error('received TLS alert record, aborting!');
-      case RECORD_TYPE.APPLICATION_DATA:
-        return await this._state.recvApplicationData(buf);
-      case RECORD_TYPE.HANDSHAKE:
-        // For simplicity, we assume that handshake messages will not be
-        // fragmented across multiple records.  They shouldn't need to be
-        // for the tiny subset of TLS that we're using.
-        do {
-          // XXX TODO: remember buffer position here.
-          const msg = readHandshakeMessage(buf);
-          // XXX TODO: read back to prev buffer position, add it to the transcript hash.
-          await this._state.recvHandshakeMessage(msg);
-        } while (buf.hasMoreBytes());
-        return null;
-      default:
-        assert(false, `unknown record type: ${type}`);
-    }
-  }
-
-  // This is a placeholder, until we implement the full key schedule.
-  // XXX TODO: the full key schedule.
-
-  _updateTrafficKeys() {
-    this._recordSender.setContentKey(this.psk, new Uint8Array(32));
-    this._recordReceiver.setContentKey(this.psk, new Uint8Array(32));
+  get channelKey() {
+    return this._channelKey;
   }
 }
-
-
-export class InsecureClientConnection extends InsecureConnection {
-  static async create(psk, pskId, sendCallback) {
-    const instance = await super.create(psk, pskId, sendCallback);
-    await instance._transition(STATE.CLIENT_START);
-    return instance;
-  }
-}
-
-
-export class InsecureServerConnection extends InsecureConnection {
-  static async create(psk, pskId, sendCallback) {
-    const instance = await super.create(psk, pskId, sendCallback);
-    await instance._transition(STATE.SERVER_START);
-    return instance;
-  }
-}
-
-
-// Re-export helpful utilities for calling code to use.
-export { bytesToHex, hexToBytes, bytesToUtf8, utf8ToBytes };
